Use podman directly & move env file to $HOME
Now that we're using --env-file we can go back to using podman
directly in mcp.json.

I was a little surprised that ~/ works, but it seems like it does.
This makes things a little less confusing since we no longer need to
care about the location of the code.

With this change, the example.mcp.json is used exactly as is - no
second yq edit required.