Just use ubuntu-latest in the GitHub workflow
I had the idea that I needed Fedora to use podman and skopeo, but
actually I'm getting errors trying to use podman inside the
container. I expect there is a solution for that, but in the short
term, let's use whatever works.

It seems like GitHub's ubuntu-latest already has the tools we need,
so no need to install podman or skopeo.